Update overall prize money, add punctuation to new categories, update overall category number, and clarify FAQ how to contribute.
Update overall prize money in the "What is it" section from $30k to $60k due to adding 3 new categories.  

Append punctuation to the three new categories in the "Prizes" section, and comma to charm maintainer wording. Also update the category number from three to six in the "Prize" section.

Clarify in the "How do I actually contribute" in the FAQ section that exports are done via the Juju GUI.